@@ -1,10 +1,11 @@
 module github.com/easysoft/zendata
 
-go 1.19
+go 1.21
 
 require (
-	github.com/getkin/kin-openapi v0.112.0
-	golang.org/x/text v0.5.0
+	github.com/asaskevich/govalidator v0.0.0-20230301143203-a9d515a09cc2
+	github.com/getkin/kin-openapi v0.120.0
+	golang.org/x/text v0.13.0
 )
 
 require (
