@@ -2,6 +2,10 @@ package controller
 
 import (
 	"bytes"
+	"io"
+	"path/filepath"
+	"strings"
+
 	consts "github.com/easysoft/zendata/internal/pkg/const"
 	"github.com/easysoft/zendata/internal/pkg/domain"
 	"github.com/easysoft/zendata/internal/pkg/service"
@@ -9,9 +13,6 @@ import (
 	logUtils "github.com/easysoft/zendata/pkg/utils/log"
 	"github.com/easysoft/zendata/pkg/utils/vari"
 	"github.com/kataras/iris/v12"
-	"io"
-	"path/filepath"
-	"strings"
 )
 
 type DataCtrl struct {
