@@ -2,7 +2,7 @@ package serverService
 
 import (
 	"github.com/easysoft/zendata/internal/pkg/model"
-	"github.com/easysoft/zendata/internal/server/repo"
+	serverRepo "github.com/easysoft/zendata/internal/server/repo"
 )
 
 type ReferService struct {
