@@ -2,15 +2,16 @@ package logUtils
 
 import (
 	"fmt"
-	"github.com/easysoft/zendata"
-	constant "github.com/easysoft/zendata/internal/pkg/const"
-	commonUtils "github.com/easysoft/zendata/pkg/utils/common"
-	"github.com/easysoft/zendata/pkg/utils/vari"
-	"github.com/fatih/color"
 	"net/http"
 	"os"
 	"regexp"
 	"strings"
+
+	zd "github.com/easysoft/zendata"
+	constant "github.com/easysoft/zendata/internal/pkg/const"
+	commonUtils "github.com/easysoft/zendata/pkg/utils/common"
+	"github.com/easysoft/zendata/pkg/utils/vari"
+	"github.com/fatih/color"
 )
 
 var (
